+++ java/vectorplugin/src/test/java/org/hansken/plugin/extraction/VectorPluginIT.java
@@ -1,0 +1,31 @@
+package org.hansken.plugin.extraction;
+
+import static nl.minvenj.nfi.flits.util.FlitsUtil.srcPath;
+
+import java.nio.file.Path;
+
+import org.hansken.plugin.extraction.api.ExtractionPlugin;
+import org.hansken.plugin.extraction.test.EmbeddedExtractionPluginFlits;
+
+class VectorPluginIT extends EmbeddedExtractionPluginFlits {
+
+    @Override
+    protected ExtractionPlugin pluginToTest() {
+        return new VectorPlugin();
+    }
+
+    @Override
+    public Path testPath() {
+        return srcPath("integration/inputs");
+    }
+
+    @Override
+    public Path resultPath() {
+        return srcPath("integration/results");
+    }
+
+    @Override
+    public boolean regenerate() {
+        return false;
+    }
+}