+++ python/fsevents/plugin/fsevents_plugin.py
@@ -764,8 +764,7 @@ class FSEventsPlugin(ExtractionPlugin):
 
     def plugin_info(self) -> PluginInfo:
         log.info('pluginInfo request')
-        plugin_info = PluginInfo(self,
-                                 id=PluginId('nfi.nl', 'events', 'FSEventsPluginPython'),
+        plugin_info = PluginInfo(id=PluginId('nfi.nl', 'events', 'FSEventsPluginPython'),
                                  version='1.0.0',
                                  description='Example Extraction Plugin: Parses apple filesystem events',
                                  author=Author('MoDS', 'mods@holmes.nl', 'NFI'),